refactor: use proper type for key arr struct array size
    • I used int as the type for the size of the
      flexible array member in the MapKeyArr struct
      where size_t would be proper.

+++ src/mmv.h
@@ -19,7 +19,7 @@ struct StrPairNode
 
 struct MapKeyArr
 {
-    int num_keys;
+    size_t num_keys;
     int keyarr[];
 };
